rearranging componentry
