+++ frontend/src/components/__tests__/DateRangeSelect.js
@@ -3,7 +3,28 @@ import React from 'react';
 import {
   render, screen, fireEvent,
 } from '@testing-library/react';
-import DateRangeSelect from '../DateRangeSelect';
+import DateRangeSelect, { formatDateRange } from '../DateRangeSelect';
+
+describe('format date function', () => {
+  it('returns a formatted date string', () => {
+    const str = formatDateRange({
+      lastThirtyDays: false,
+      string: '2021/06/07-2021/06/08',
+      withSpaces: true,
+    });
+
+    expect(str).toBe('06/07/2021 - 06/08/2021');
+  });
+
+  it('returns a formatted date string without spaces', () => {
+    const str = formatDateRange({
+      string: '2021/06/07-2021/06/08',
+      withSpaces: false,
+    });
+
+    expect(str).toBe('06/07/2021-06/08/2021');
+  });
+});
 
 describe('DateRangeSelect', () => {
   const renderDateRangeSelect = (onApplyDateRange) => {
