+++ frontend/src/components/constants.js
@@ -4,10 +4,12 @@ export const BEFORE = 'Is before';
 export const AFTER = 'Is after';
 export const WITHIN = 'Is within';
 export const IS = 'Is';
+export const IS_NOT = 'Is not';
 export const ONE_OF = 'One of';
 
 export const SELECT_CONDITIONS = [CONTAINS, NOT_CONTAINS];
 export const DATE_CONDITIONS = [BEFORE, AFTER, WITHIN];
+export const FILTER_CONDITIONS = [IS, IS_NOT];
 
 export const QUERY_CONDITIONS = {
   [CONTAINS]: 'in[]',
@@ -16,4 +18,22 @@ export const QUERY_CONDITIONS = {
   [AFTER]: 'aft',
   [WITHIN]: 'win',
   [IS]: 'is',
+  [IS_NOT]: 'not',
 };
+
+export const DATE_FORMAT = 'MM/DD/YYYY';
+export const DATETIME_DATE_FORMAT = 'YYYY/MM/DD';
+
+export const DATE_OPTIONS = [
+  {
+    label: 'Last 30 Days',
+    value: 1,
+  },
+  {
+    label: 'Custom Date Range',
+    value: 2,
+  },
+];
+
+export const LAST_THIRTY_DAYS = DATE_OPTIONS[0].value;
+export const CUSTOM_DATE_RANGE = DATE_OPTIONS[1].value;
