+++ frontend/src/pages/GranteeRecord/pages/TTAHistory.js
@@ -3,10 +3,10 @@ import PropTypes from 'prop-types';
 import Overview from '../../../widgets/DashboardOverview';
 import FilterMenu from '../../../components/FilterMenu';
 
-export default function TTAHistory({ filters }) {
+export default function TTAHistory({ filters, onApplyFilters }) {
   return (
     <div className="margin-right-3">
-      <FilterMenu filters={filters} />
+      <FilterMenu filters={filters} onApplyFilters={onApplyFilters} />
       <Overview filters={filters} />
     </div>
   );
@@ -19,6 +19,7 @@ TTAHistory.propTypes = {
     condition: PropTypes.string,
     query: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
   })),
+  onApplyFilters: PropTypes.func.isRequired,
 };
 
 TTAHistory.defaultProps = {
