--- frontend/src/pages/RegionalDashboard/constants.js
@@ -1,16 +1,0 @@
-export const DATE_FORMAT = 'MM/DD/YYYY';
-export const DATETIME_DATE_FORMAT = 'YYYY/MM/DD';
-
-export const DATE_OPTIONS = [
-  {
-    label: 'Last 30 Days',
-    value: 1,
-  },
-  {
-    label: 'Custom Date Range',
-    value: 2,
-  },
-];
-
-export const LAST_THIRTY_DAYS = DATE_OPTIONS[0].value;
-export const CUSTOM_DATE_RANGE = DATE_OPTIONS[1].value;
